backend/pyproject.toml
[project]
name = "aya-shield-backend"
version = "0.1.0"
description = "Aya Shield — AI-Powered Transaction Firewall & Smart Receipt API"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Suyash Kumar Singh" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.8.4",
    "web3>=7.6.0",
    "httpx>=0.28.0",
    "supabase>=2.11.0",
    "redis>=5.2.0",
    "python-jose[cryptography]>=3.3.0",
    "slowapi>=0.1.9",
    "eth-abi>=5.1.0",
    "email-validator>=2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
